pyproject.toml
[tool.poetry]
name = "phenoseeker"
version = "0.1.0"
description = ""
authors = ["Maxime Sanchez <maxime.sanchez14@gmail.com>"]
license = "Creative Commons Attribution-NonCommercial 4.0 International (CC BY-NC 4.0)"
readme = "README.md"
packages = [
    { include = "phenoseeker", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
numpy = "^2.3.4"
pandas = "^2.3.3"
tqdm = "^4.67.1"
torch = "^2.9.0"
torchvision = "^0.24.0"
psutil = "^7.1.3"
pillow = "^12.0.0"
pyyaml = "^6.0.3"
scipy = "^1.16.3"
joblib = "^1.5.2"
matplotlib = "^3.10.7"
tqdm-joblib = "^0.0.5"
seaborn = "^0.13.2"
scikit-learn = "^1.7.2"
umap-learn = "^0.5.9.post2"
rdkit = "^2025.9.1"
pingouin = "^0.5.5"
datasketch = "^1.7.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
